reorganize exporter and add multipatch
