+++ cpp/splinepy/py/py_spline.hpp
@@ -901,7 +901,6 @@ public:
   }
 };
 
-
 inline void add_spline_pyclass(py::module& m) {
   py::class_<splinepy::py::PySpline, std::shared_ptr<splinepy::py::PySpline>>
       klasse(m, "CoreSpline");
