+++ cpp/splinepy/py/splinepy_core.cpp
@@ -29,6 +29,9 @@ void init_fitting(py::module_&);
 // unique vertices
 void init_uffpy(py::module_&);
 
+// multi_patch
+void init_multi_patch(py::module_& m);
+
 } // namespace splinepy::py::init
 
 namespace py = pybind11;
@@ -44,4 +47,5 @@ PYBIND11_MODULE(splinepy_core, m) {
   splinepy::py::init::init_exporter(m);
   splinepy::py::init::init_fitting(m);
   splinepy::py::init::init_uffpy(m);
+  splinepy::py::init::init_multi_patch(m);
 }
